web-apps/reports/requirements.txt
#
# This file is autogenerated by pip-compile
# To update, run:
#
#    pip-compile
#
cachecontrol==0.12.6
    # via firebase-admin
cachetools==4.2.2
    # via google-auth
certifi==2020.12.5
    # via requests
cffi==1.14.5
    # via google-crc32c
chardet==4.0.0
    # via requests
click==7.1.2
    # via uvicorn
fastapi==0.63.0
    # via -r requirements.in
firebase-admin==5.0.0
    # via -r requirements.in
google-api-core[grpc]==1.26.3
    # via
    #   firebase-admin
    #   google-api-python-client
    #   google-cloud-bigquery
    #   google-cloud-core
    #   google-cloud-error-reporting
    #   google-cloud-firestore
    #   google-cloud-logging
google-api-python-client==2.4.0
    # via firebase-admin
google-auth-httplib2==0.1.0
    # via google-api-python-client
google-auth==1.30.0
    # via
    #   google-api-core
    #   google-api-python-client
    #   google-auth-httplib2
    #   google-cloud-core
    #   google-cloud-storage
google-cloud-bigquery==2.16.1
    # via -r requirements.in
google-cloud-core==1.6.0
    # via
    #   google-cloud-bigquery
    #   google-cloud-firestore
    #   google-cloud-logging
    #   google-cloud-storage
google-cloud-error-reporting==1.1.2
    # via -r requirements.in
google-cloud-firestore==2.1.1
    # via firebase-admin
google-cloud-logging==2.3.1
    # via google-cloud-error-reporting
google-cloud-storage==1.38.0
    # via firebase-admin
google-crc32c==1.1.2
    # via google-resumable-media
google-resumable-media==1.2.0
    # via
    #   google-cloud-bigquery
    #   google-cloud-storage
googleapis-common-protos==1.53.0
    # via google-api-core
grpcio==1.37.1
    # via google-api-core
h11==0.12.0
    # via uvicorn
httplib2==0.19.1
    # via
    #   google-api-python-client
    #   google-auth-httplib2
httptools==0.1.2
    # via uvicorn
idna==2.10
    # via requests
joblib==1.0.1
    # via scikit-learn
msgpack==1.0.2
    # via cachecontrol
numpy==1.20.3
    # via
    #   pandas
    #   scikit-learn
    #   scipy
packaging==20.9
    # via
    #   google-api-core
    #   google-cloud-bigquery
pandas==1.2.4
    # via -r requirements.in
proto-plus==1.18.1
    # via
    #   google-cloud-bigquery
    #   google-cloud-error-reporting
    #   google-cloud-firestore
    #   google-cloud-logging
protobuf==3.17.0
    # via
    #   google-api-core
    #   google-cloud-bigquery
    #   googleapis-common-protos
    #   proto-plus
pyasn1-modules==0.2.8
    # via google-auth
pyasn1==0.4.8
    # via
    #   pyasn1-modules
    #   rsa
pycparser==2.20
    # via cffi
pydantic==1.8.2
    # via fastapi
pyparsing==2.4.7
    # via
    #   httplib2
    #   packaging
python-dateutil==2.8.1
    # via pandas
python-dotenv==0.17.0
    # via uvicorn
pytz==2021.1
    # via
    #   google-api-core
    #   google-cloud-firestore
    #   pandas
pyyaml==5.4.1
    # via uvicorn
requests==2.25.1
    # via
    #   cachecontrol
    #   google-api-core
    #   google-cloud-bigquery
    #   google-cloud-storage
rsa==4.7.2
    # via google-auth
scikit-learn==0.24.2
    # via -r requirements.in
scipy==1.6.3
    # via scikit-learn
six==1.16.0
    # via
    #   google-api-core
    #   google-api-python-client
    #   google-auth
    #   google-auth-httplib2
    #   google-cloud-core
    #   google-resumable-media
    #   grpcio
    #   protobuf
    #   python-dateutil
starlette==0.13.6
    # via fastapi
threadpoolctl==2.1.0
    # via scikit-learn
typing-extensions==3.7.4.3
    # via pydantic
uritemplate==3.0.1
    # via google-api-python-client
urllib3==1.26.4
    # via requests
uvicorn[standard]==0.13.4
    # via -r requirements.in
uvloop==0.15.2
    # via uvicorn
watchgod==0.7
    # via uvicorn
websockets==8.1
    # via uvicorn

# The following packages are considered to be unsafe in a requirements file:
# setuptools
